% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfun.R
\name{myfirstfun}
\alias{myfirstfun}
\title{My first R function}
\usage{
myfirstfun(x)
}
\arguments{
\item{x}{A vector of quantitatve data}
}
\value{
A vector of squared components
}
\description{
This is a first study on R functions
}
\details{
The course MATH 4753 is an introduction to univariate stats with a heavy influence on tools and computer applications of statistics
}
\examples{
y <- 1:10; myfirstfun(y)
}
